Delete CardView and Add UIImageView Extension. #6
